Use eml_version() to set version
This approach allows us to have a single point that needs to be changed to reset the default version of EML for the package, instead of needing to change the fallback value in `getOption` everywhere it occurs!

This also puts the default back to the current 2.1.1 while 2.2.0 is still officially not yet released.