Fix a bug where TextType content got newlines it shouldn't have
@jeanetteclark and others found this while adding subscripts to
an EML document. What happens is that as_jsonlist is reading in
the contents of TextType `para` and `section` elements, and
converting them to literal XML strings with this code.

<para>H<subscript>2</subscript>O</para> gets turned into

para
  H
  <subscript>2</subscript>
  O

and calling as.character on the above followed by a paste with
collapse = "\n" introduces newlines between each child of the para.
This isn't a huge issue for most uses cases but, when rendering to
HTML via XSLT, you end up with spaces between the H and the 2
because browsers are getting the the newline and converting it to
whitespace.

Fixes https://github.com/ropensci/EML/issues/282