Make validate_units use the right eml-unitDictionary
Ref #56

While `eml_validate` determines the root schema to use when validating, `validate_units` was still defaulting to the version set in `emld_db`. This changes `validate_units` so it uses the appropriate `eml-unitDictionary.xml` file for the root element.

I've also added two test files and a test to exercise this so we know validate_units is working correctly now.